refactor: backspace and space buttons in moud and keyboard handler
